Removed var init on channel data
There was an issue that the rest of the variables that aren't assigned to var are printed while these initialized aren't.

It's kinda weird that this was working in the past but now it wasn't. Dunno if its a browser update change or guess what? So far, things got better without it.